init new asp.net mvc project (#2)
* feat: Add base entity interface and infrastructure class

This commit adds the `IBaseEntity` interface to the `Dev.Core` namespace and the `Class1` class to the `Dev.Infrastructure` namespace. These changes lay the foundation for future development by providing a base entity interface and a sample infrastructure class.

Note: The commit message has been generated based on the provided code changes and recent commits.

* Add Dev.Plugin.Blog project to solution

* feat: Add HomeController and views for home page

This commit adds the HomeController and corresponding views for the home page. The HomeController is responsible for rendering the home page view, which displays the "Home page" heading. This change improves the user experience by providing a basic home page for the application.

Note: The commit message has been generated based on the provided code changes and recent commits.