refactor: Update navbar branding and add Blog link
This commit updates the navbar branding in the _Layout.cshtml file, changing the text from "WebApp" to "devbetter". Additionally, it adds a new Blog link to the navbar, pointing to the Blog home page. These changes improve the branding and navigation of the website.

Note: The commit message has been generated based on the provided code changes and recent commits.